pyproject.toml
[tool.poetry]
name = "selenium-python"
version = "0.1.0"
description = ""
authors = ["None"]
readme = "README.md"
packages = [{include = "selenium_python"}]

[tool.poetry.dependencies]
python = "^3.10"
selenium = "^4.7.2"
fake-useragent = "^1.1.1"
pymysql = "^1.0.2"
python-dotenv = "^0.21.0"
loguru = "^0.6.0"
psycopg2-binary = "^2.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
